settings: Extract computed settings to computed_settings.
